Encapsulate message printing in a function
Case CPANEL-20161: The use of echo -n is not portable according to
POSIX, and not all shells support it.  All but one of these uses are in
printing user-visible messages.  Encapsulate the printing of these
messages in a function, and switch to printf, which is portable.  Since
we are using printf, avoid interpolating variables into the first
argument so that we don't print any lingering escape sequences we may
not have intended.